Resolve circular dependency in StudyViewPageStore
