+++ src/pages/studyView/addChartButton/addChartByType/AddChartByType.tsx
@@ -9,7 +9,7 @@ import {Column} from "../../../../shared/components/lazyMobXTable/LazyMobXTable"
 import {getFrequencyStr} from "../../StudyViewUtils";
 import LoadingIndicator from "../../../../shared/components/loadingIndicator/LoadingIndicator";
 import MobxPromise from 'mobxpromise';
-import {ClinicalDataCountSet} from "../../StudyViewPageStore";
+import {ClinicalDataCountSet} from "../../StudyViewUtils";
 import FixedHeaderTable from "../../table/FixedHeaderTable";
 import autobind from 'autobind-decorator';
 import classnames from 'classnames';
