+++ src/pages/studyView/addChartButton/customCaseSelection/CustomCaseSelection.tsx
@@ -5,7 +5,7 @@ import {observer} from "mobx-react";
 import styles from "./styles.module.scss";
 import {action, computed, observable} from 'mobx';
 import {ButtonGroup, Modal, Radio} from 'react-bootstrap';
-import {ClinicalDataType, ClinicalDataTypeEnum, NewChart} from "../../StudyViewPageStore";
+import {NewChart} from "../../StudyViewPageStore";
 import ErrorBox from "../../../../shared/components/errorBox/ErrorBox";
 import {STUDY_VIEW_CONFIG} from "../../StudyViewConfig";
 import {
@@ -17,6 +17,7 @@ import {
 import autobind from 'autobind-decorator';
 import Collapse from "react-collapse";
 import {serializeEvent} from "../../../../shared/lib/tracking";
+import {ClinicalDataType, ClinicalDataTypeEnum} from "../../StudyViewUtils";
 
 export interface ICustomCaseSelectionProps {
     allSamples: Sample[];
