+++ src/pages/studyView/charts/scatterPlot/StudyViewScatterPlot.tsx
@@ -13,7 +13,7 @@ import {DOWNSAMPLE_PIXEL_DISTANCE_THRESHOLD, getDownsampledData, MAX_DOT_SIZE} f
 import {ClinicalAttribute, SampleIdentifier} from "../../../../shared/api/generated/CBioPortalAPI";
 import LoadingIndicator from "shared/components/loadingIndicator/LoadingIndicator"
 import $ from "jquery";
-import {AnalysisGroup} from "../../StudyViewPageStore";
+import {AnalysisGroup} from "../../StudyViewUtils";
 import {AbstractChart} from "../ChartContainer";
 
 export interface IStudyViewScatterPlotData {
