+++ src/pages/studyView/charts/survival/StudyViewSurvivalUtils.ts
@@ -1,6 +1,6 @@
 import {PatientSurvival} from "../../../../shared/model/PatientSurvival";
 import {IChartContainerProps} from "../ChartContainer";
-import {AnalysisGroup} from "../../StudyViewPageStore";
+import {AnalysisGroup} from "../../StudyViewUtils";
 import _ from "lodash";
 
 export function makeSurvivalChartData(
